+++ src/Types/isBool.ts
@@ -1,0 +1,11 @@
+/**
+ * Created by Martin Neundorfer on 01.02.2019.
+ * For LABOR.digital
+ */
+/**
+ * Returns true if the given value is a boolean value
+ * @param value
+ */
+export function isBool(value):boolean {
+	return value === true || value === false;
+}