+++ src/Ui/scrollToPosition.ts
@@ -3,19 +3,14 @@
  * For LABOR.digital
  */
 import {throttleEvent} from "./throttleEvent";
+import {requestFrame} from "../Browser/requestFrame";
+import {getScrollPos} from "./getScrollPos";
 
 // The number of milliseconds for each frame / tick
 const tickLength = 15;
 
 // The id of the running animation, or zero
 let runningAnimation = 0;
-
-/**
- * Polyfill for request animation frame
- */
-const requestFrame = window.requestAnimationFrame.bind(window) || function (c) {
-	setTimeout(c, 1000 / 60)
-};
 
 /**
  * The used easing function
@@ -32,7 +27,7 @@ function easing(p) {
  * @param duration (Default 300) The duration in milliseconds the animation should take
  * @param container (Default window) The container to scroll instead of the window
  */
-export function scrollToPosition(position: number, duration?: number, container?: HTMLElement|Window): Promise<HTMLElement|Window> {
+export function scrollToPosition(position: number, duration?: number, container?: HTMLElement | Window): Promise<HTMLElement | Window> {
 	// Make sure we break if a new animation was started
 	const localAnimation = Math.random();
 	runningAnimation = localAnimation;
@@ -47,45 +42,41 @@ export function scrollToPosition(position: number, duration?: number, container?
 		const ticks = Math.floor(duration / tickLength);
 		let c = 1;
 
-		// Helper to get the scroll value
-		const getScrollPos = function () {
-			if(containerIsWindow) return window.scrollY || window.pageYOffset || document.documentElement.scrollTop;
-			return (container as HTMLElement).scrollTop;
-		};
 
 		// Helper to set the scroll value
-		const setScrollPos = function(pos){
-			if(containerIsWindow) (container as Window).scrollTo(0, pos);
+		const setScrollPos = function (pos) {
+			if (containerIsWindow) (container as Window).scrollTo(0, pos);
 			else (container as HTMLElement).scrollTop = pos;
 		};
 
 		// Prepare calculation
-		const initialPosition = getScrollPos();
+		const initialPosition = getScrollPos(container as HTMLElement);
 		const distance = position - initialPosition;
 
 		// Duration is zero -> No animation
-		if(duration === 0){
+		if (duration === 0) {
 			setScrollPos(position);
 			resolve(container);
 		}
 
 		// Nothing to do
-		if(distance === 0) return resolve(container);
+		if (distance === 0) return resolve(container);
 
 		// Marker to detect if the user changed our expected scroll position
-		let expectedPosition = initialPosition;
+		let expectedPosition = Math.ceil(initialPosition);
 
 		// The animation loop
 		const tick = function () {
 			// Break if another animation was started or something else scrolled
-			if(localAnimation !== runningAnimation ||
-				expectedPosition !== getScrollPos()) return resolve(container);
+			if (localAnimation !== runningAnimation || (expectedPosition !== Math.ceil(getScrollPos(container as HTMLElement)))) {
+				return resolve(container);
+			}
 
 			// Check if there is still work to do
 			if (c < ticks) {
-				const p = easing(c /ticks);
+				const p = easing(c / ticks);
 				requestFrame(throttleEvent(tick, tickLength) as any);
-				expectedPosition = Math.round(initialPosition + (distance * p));
+				expectedPosition = Math.ceil(initialPosition + (distance * p));
 				setScrollPos(expectedPosition);
 			} else {
 				setScrollPos(position);